move activty-role-related functionality into dedicated room extension
