+++ lib/pangea/chat_settings/utils/delete_room_extension.dart
@@ -4,6 +4,7 @@ import 'package:http/http.dart';
 import 'package:matrix/matrix.dart';
 import 'package:matrix/matrix_api_lite/generated/api.dart';
 
+import 'package:fluffychat/pangea/activity_sessions/activity_roles_room_extension.dart';
 import 'package:fluffychat/pangea/activity_sessions/activity_room_extension.dart';
 import 'package:fluffychat/pangea/events/constants/pangea_room_types.dart';
